+++ examples/resources/artifactkeeper_totp_policy/import.sh
@@ -1,0 +1,1 @@
+terraform import artifactkeeper_totp_policy.this totp_policy